+++ GameVote.Domain/Entities/Home/Game.cs
@@ -3,7 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace GameVote.Domain.Entities
+namespace GameVote.Domain.Entities.Home
 {
     public class Game : IGame
     {
@@ -12,5 +12,7 @@ namespace GameVote.Domain.Entities
         public string Opisanie { get; set; }
         public float PriceRelise { get; set; }
         public float PriceNow { get; set; }
+        public string UrlGame { get; set; }
+        public string[] ImgGame { get; set; }
     }
 }
