+++ GameVote.Domain/Entities/Home/Interface/IGame.cs
@@ -4,13 +4,14 @@ using System.Text;
 
 namespace GameVote.Domain.Entities.Home.Interface
 {
-    interface IGame
+    public interface IGame
     {
         ulong Id { get; set; }
         string Nazvanie { get; set; }
         string Opisanie { get; set; }
         float PriceRelise { get; set; }
         float PriceNow { get; set; }
-
+        string UrlGame { get; set; }
+        string[] ImgGame { get; set; }
     }
 }
