+++ GameVote/Views/Home/Index.cshtml
@@ -1,0 +1,16 @@
+﻿<head>
+    <meta charset="utf-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <title>НАЧАЛО РАБОТЫ С BRACKETS</title>
+    <meta name="description" content="Интерактивное руководство по началу работы в Brackets.">
+    <link rel="stylesheet" href="main.css">
+</head>
+<section>
+    <div class="header">
+        <img class="logo">
+        <a href="oproekte">О проекте</a>
+        <a href="index.html">Главная</a>
+        <button class="login" value="Войти" style="font-size: 100px 40px 100px 100px" />
+        <text></text>
+    </div>
+</section>
